requirements.txt
Flask==0.10.1
Flask-Assets==0.8
Flask-Login==0.2.7
Flask-Mail==0.9.0
Flask-OAuth==0.12
Flask-SQLAlchemy==0.16
Flask-WTF==0.9.3
Jinja2==2.7.1
MarkupSafe==0.18
SQLAlchemy==0.7.9
Tempita==0.5.1
WTForms==1.0.5
Werkzeug==0.9.4
argparse==1.2.1
beautifulsoup4==4.3.2
blinker==1.3
chartkick==0.4.0
decorator==3.4.0
enum34==0.9.19
feedparser==5.1.3
httplib2==0.8
itsdangerous==0.23
nfldb==0.0.11
nflgame==1.1.21
oauth2==1.5.211
psycopg2==2.5.1
python-crontab==1.5.1
python-dateutil==2.2
pytz==2013b
six==1.4.1
sqlalchemy-migrate==0.7.2
stevedore==0.12
virtualenv==1.10.1
virtualenv-clone==0.2.4
webassets==0.8
wsgiref==0.1.2
